add test for GET & POST endpoints!
